# load category stats only once, remove bottom and top 5% before comput…