…ing mean and std, bug #2436

I just remarked that we were loading a 10 Mb file with category stats for each product page view... I'm now loading them at apache startup. They are updated only nightly, and they don't change that much, so it's ok to reload them only when we restart the server.