@hangy 2019-10-15
Did you have a look at the additional memory usage after this change? From my understanding, OFF already uses quite some memory in production, and adding loading this for every country/category combination could result in huge additional memory usage. I suppose it's not 10 MiB for every country/category combination (that would be >70 TiB of RAM 😆), but it could be substantial nonetheless.