@stephanegigandet 2019-10-16
@hangy : It's 10 Mb for the world file that contains stats for all categories. The total for all countries including world is less than 50 Mb.

But today if you load a product page on world, each time it loads the 10 Mb world file. That does not make sense.