@stephanegigandet 2019-10-16
I did some performance testing, on the dev server, using only one apache process. (serverlimit set to 1).
Just requesting the same product page from world is 4 times faster when the category statistics are preloaded at the start of the module:

Without preloading:

root@ns3135968:/home/off/openfoodfacts-server# ab -k -c 1 -A off:off -n 100 https://world.openfoodfacts.dev/product/3228857000661/mini-brioches-tressees-harrys
This is ApacheBench, Version 2.3 <$Revision: 1757674 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking world.openfoodfacts.dev (be patient).....done


Server Software:        nginx/1.10.3
Server Hostname:        world.openfoodfacts.dev
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
TLS Server Name:        world.openfoodfacts.dev

Document Path:          /product/3228857000661/mini-brioches-tressees-harrys
Document Length:        68761 bytes

Concurrency Level:      1
Time taken for tests:   40.726 seconds
Complete requests:      100
Failed requests:        0
Keep-Alive requests:    0
Total transferred:      6929700 bytes
HTML transferred:       6876100 bytes
Requests per second:    2.46 [#/sec] (mean)
Time per request:       407.255 [ms] (mean)
Time per request:       407.255 [ms] (mean, across all concurrent requests)
Transfer rate:          166.17 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        9    9   1.2      9      20
Processing:   393  398   7.2    396     453
Waiting:      392  397   7.2    395     452
Total:        402  407   8.2    405     473

Percentage of the requests served within a certain time (ms)
  50%    405
  66%    406
  75%    407
  80%    407
  90%    415
  95%    417
  98%    433
  99%    473
 100%    473 (longest request)
 
With preloading:
 
root@ns3135968:/home/off/openfoodfacts-server# ab -k -c 1 -A off:off -n 100 https://world.openfoodfacts.dev/product/3228857000661/mini-brioches-tressees-harrys
This is ApacheBench, Version 2.3 <$Revision: 1757674 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking world.openfoodfacts.dev (be patient).....done


Server Software:        nginx/1.10.3
Server Hostname:        world.openfoodfacts.dev
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
TLS Server Name:        world.openfoodfacts.dev

Document Path:          /product/3228857000661/mini-brioches-tressees-harrys
Document Length:        68761 bytes

Concurrency Level:      1
Time taken for tests:   10.886 seconds
Complete requests:      100
Failed requests:        0
Keep-Alive requests:    0
Total transferred:      6929700 bytes
HTML transferred:       6876100 bytes
Requests per second:    9.19 [#/sec] (mean)
Time per request:       108.862 [ms] (mean)
Time per request:       108.862 [ms] (mean, across all concurrent requests)
Transfer rate:          621.64 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        9    9   1.2      9      20
Processing:    96  100  10.6     98     203
Waiting:       95   99  10.5     97     200
Total:        105  109  11.7    107     222

Percentage of the requests served within a certain time (ms)
  50%    107
  66%    108
  75%    108
  80%    108
  90%    110
  95%    115
  98%    116
  99%    222
 100%    222 (longest request)