@stephanegigandet 2019-10-16
In terms of memory usage, after putting back multiple apache processes, and doing 100 requests to the same product with 10 concurrent requests, the memory of the apache processes is in fact smaller after (don't ask me why):

Without preloading:
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 28169
28169 off       20   0  953228 665772   7916 S   0.0  2.0   0:01.62 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 28171
28171 off       20   0  953228 665772   7916 S   0.0  2.0   0:01.08 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 28172
28172 off       20   0  953228 665776   7916 S   0.0  2.0   0:01.09 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 28173
28173 off       20   0  953228 665772   7916 S   0.0  2.0   0:01.09 /usr/sbin/+


With preloading:

27936 off       20   0  946668 661444   7864 S   0.0  2.0   0:01.65 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 27939
27939 off       20   0  946532 661408   7864 S   0.0  2.0   0:00.57 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 27923
27923 off       20   0  946676 660736   7020 S   0.0  2.0   0:02.62 /usr/sbin/+
root@ns3135968:/home/off/openfoodfacts-server# top -b -n 1 | grep 27924
27924 off       20   0  946676 660708   6992 S   0.0  2.0   0:02.59 /usr/sbin/+