@stephanegigandet 2019-10-16
With multiple apache processes and concurrent requests:

Without preloading: 9 requests per sec.

With preloading: 28 requests per sec.